+++ CHANGES.md
@@ -2,6 +2,7 @@
 =====
 * Implement low-level audio manipulation in C
   to avoid needless allocations.
+* Add Audio.Mono.squares
 
 0.7.3 (10-12-2021)
 =====